Feature/158 delete project sprint command new business rule there has to be at least one sprint

+++ src/ConsiliumTempus.Application/ProjectSprint/Commands/Delete/DeleteProjectSprintCommandHandler.cs
@@ -13,15 +13,17 @@ public sealed class DeleteProjectSprintCommandHandler(
     public async Task<ErrorOr<DeleteProjectSprintResult>> Handle(DeleteProjectSprintCommand command,
         CancellationToken cancellationToken)
     {
-        var projectSprint = await projectSprintRepository.GetWithWorkspace(
+        var projectSprint = await projectSprintRepository.GetWithSprintsAndWorkspace(
             ProjectSprintId.Create(command.Id),
             cancellationToken);
+
         if (projectSprint is null) return Errors.ProjectSprint.NotFound;
+        if (projectSprint.Project.Sprints.Count == 1) return Errors.ProjectSprint.OnlyOneSprint;
 
         projectSprintRepository.Remove(projectSprint);
 
         projectSprint.Project.RefreshActivity();
-        
+
         return new DeleteProjectSprintResult();
     }
 }
